% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSopiCheck.R
\name{cCheck}
\alias{cCheck}
\title{Identify conflicts between two categories in a table}
\usage{
cCheck(.data = comtrader::omtcodes, ref1, ref2)
}
\arguments{
\item{.data}{a table to read data from. Defaults to `omtcodes`.}

\item{ref1}{first reference column.}

\item{ref2}{second reference column.}
}
\value{
a tibble.
}
\description{
Creates a table of counts of matching categories between two columns of a table.
}
\details{
the table is built with reference to the `ref1` argument.
}
\examples{
cCheck(ref1 = NZHSCLevel4, ref2 = `SOPI_group_HS6`)
}
