% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodityRecode.R
\name{commodityRecode}
\alias{commodityRecode}
\title{Create named vector for HS codes}
\usage{
commodityRecode(
  hsLevel = 6,
  sopiLevel = `SOPI Forecast Group`,
  sopiFilter = NULL,
  query = TRUE
)
}
\arguments{
\item{hsLevel}{NZHSC level code (either 2, 4 or 6)}

\item{sopiLevel}{SOPI category level (either `Primary Industry Sector`,`SOPI Forecast Group` or `Granular Group`)}

\item{query}{If `TRUE` returns a vector of commodity codes (`cmdCode`) to use in a query. If `FALSE` returns a vector of `sopiLevel` names associated with the commodity code.}
}
\value{
a named vector
}
\description{
Creates a named vector matching HS codes and SOPI Categories
}
\details{
Data was extracted from the Power BI OMT dataset. This dataset needs to be updated if there are changes in the OMT classification.
}
\examples{
commodityRecode(hsLevel = 4, sopiLevel= "SOPI Forecast Group")
}
