% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryshiny.R
\docType{data}
\name{countryshiny}
\alias{countryshiny}
\title{countryshiny}
\format{
\describe{
  \item{List name}{country description.}
  \item{List item}{country code}
}
}
\source{
ARIA
}
\usage{
countryshiny
}
\description{
Named list of country codes to use in comtrader shiny app.
}
\details{
The list includes a "World" code in addition to M49 codes for individual countries.
}
\keyword{datasets}
