% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCTSopi.R
\name{getCTSopi}
\alias{getCTSopi}
\title{Custom get data from UN Comtrade Database API}
\usage{
getCTSopi(
  typeCode = "C",
  freqCode = "M",
  clCode = "HS",
  reporterCode = NULL,
  startDate = NULL,
  endDate = NULL,
  partnerCode = NULL,
  partner2Code = NULL,
  flowCode = NULL,
  customsCode = NULL,
  motCode = NULL,
  aggregateBy = NULL,
  breakdownMode = "classic",
  includeDesc = TRUE,
  sopiLevel = Primary_Industry_Sector,
  sopiFilter = NULL,
  hs = NZHSCLevel6
)
}
\arguments{
\item{typeCode}{a character string indicating type of trade: "C" for commodities and "S" for service}

\item{freqCode}{a character string indicating trade frequency: "A" for annual and "M" for monthly}

\item{clCode}{a character string indicating trade classification (IMTS): "HS", "SITC", "BEC" or "EBOPS"}

\item{reporterCode}{a character string indicating reporter code (Possible values are M49 code of the countries)}

\item{startDate}{a character string indicating start date for query: "YYYY-MM-DD" (optional).}

\item{endDate}{Optional end date for query: "YYYY-MM-DD" (optional).}

\item{partnerCode}{Partner code (M49 International Country Classification).}

\item{partner2Code}{Second partner/consignment code (M49 International Country Classification).}

\item{flowCode}{a character string indicating trade flow code:  "X" for exports, "RX" for re-exports, "M" for imports, "RM" for re-imports. Use character vector for multiple trade flow entries}

\item{customsCode}{a character string indicating customs code. Use character vector for multiple customs code entries (defaults to all)}

\item{motCode}{a character string indicating mode of transport code. Use character vector for multiple customs code entries transport code (defaults to all)}

\item{aggregateBy}{Add parameters in the form of a character vector on which you want the results to be aggregated}

\item{breakdownMode}{a character string indicating breakdown mode: "classic" (trade by partner/product: dafault) or "plus" (extended breakdown)}

\item{includeDesc}{boolean indicating if categories descriptions shoould be returned (defaults to `TRUE`)}

\item{sopiLevel}{SOPI level column in `omtcodes` (either `Primary_Industry_Sector` or `SOPI_group_HS6`)}

\item{sopiFilter}{character string to filter SOPI level (e.g.: "Dairy")}

\item{hs}{HS code level column in `omtcodes` (e.g.: `NZHSCLevel6`)}
}
\value{
a tibble
}
\description{
Custom get function to query data from the UN Comtrade API
}
\details{
for further details visit the [UN Comtrade API developer site](`https://comtradedeveloper.un.org`)
}
\examples{
getCTSopi(reporterCode = 36,
          startDate = "2020-01-01",
          endDate = "2020-02-01",
          sopiLevel = Primary_Industry_Sector,
          sopiFilter = "Dairy")
}
