% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDaTariffline.R
\name{getDaTariffline}
\alias{getDaTariffline}
\title{Get tariff line data availability from UN Comtrade Database API}
\usage{
getDaTariffline(
  typeCode = "C",
  freqCode = "M",
  clCode = "HS",
  reporterCode = NULL,
  period = NULL,
  publishedDateFrom,
  publishedDateTo
)
}
\arguments{
\item{typeCode}{Type of trade: C for commodities and S for service}

\item{freqCode}{Trade frequency: A for annual and M for monthly}

\item{clCode}{Trade (IMTS) classifications: HS, SITC, BEC or EBOPS.}

\item{reporterCode}{Reporter code (Possible values are M49 code of the countries separated by comma (,))}

\item{period}{Year or month. Year should be 4 digit year. Month should be six digit integer with the values of the form YYYYMM. Ex: 201002 for 2010 February. Multi value input should be in the form of csv (Codes separated by comma (,))}

\item{publishedDateFrom}{Publication date From YYYY-MM-DD}

\item{publishedDateTo}{Publication date To YYYY-MM-DD}
}
\value{
a tibble
}
\description{
Query tariff line data availability in the UN Comtrade API
}
\details{
further details on API features available at: `https://comtradedeveloper.un.org/api-details#api=comtrade-v1`
}
\examples{
getDaTariffline(clCode = "H6",publishedDateFrom = "2020-01-01")
}
