% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLiveUpdate.R
\name{getLiveUpdate}
\alias{getLiveUpdate}
\title{Get live update data from UN Comtrade Database API}
\usage{
getLiveUpdate()
}
\value{
a tibble.
}
\description{
Query progress on data release in the UN Comtrade API.
}
\details{
further details on API features available at: `https://comtradedeveloper.un.org`.
}
\examples{
getLiveUpdate()
}
