% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUNComtrade.R
\name{getUNComtrade}
\alias{getUNComtrade}
\title{Custom get data from UN Comtrade Database API}
\usage{
getUNComtrade(
  typeCode = "C",
  freqCode = "M",
  clCode = "HS",
  reporterCode = NULL,
  startDate = NULL,
  endDate = NULL,
  partnerCode = NULL,
  partner2Code = NULL,
  cmdCode = NULL,
  flowCode = NULL,
  customsCode = NULL,
  motCode = NULL,
  aggregateBy = NULL,
  breakdownMode = "classic",
  includeDesc = TRUE
)
}
\arguments{
\item{typeCode}{Type of trade: C for commodities and S for service}

\item{freqCode}{Trade frequency: A for annual and M for monthly}

\item{clCode}{Trade (IMTS) classifications: HS, SITC, BEC or EBOPS.}

\item{reporterCode}{Reporter code (Possible values are M49 code of the countries)}

\item{startDate}{Optional start date for query (e.g. "2020-01-01"). Multi value input should be in the form of a character vector.}

\item{endDate}{Optional end date for query (e.g. "2022-01-01").}

\item{partnerCode}{Partner code (Possible values are M49 code of the countries)}

\item{partner2Code}{Second partner/consignment code (Possible values are M49 code of the countries)}

\item{cmdCode}{Commodity code. Multi value input should be in the form of a character vector.}

\item{flowCode}{Trade flow code. Multi value input should be in the form of a character vector)}

\item{customsCode}{Customs code. Multi value input should be in the form of a character vector)}

\item{motCode}{Mode of transport code. Multi value input should be in the form of a character vector)}

\item{aggregateBy}{Add parameters in the form of a character vector on which you want the results to be aggregated}

\item{breakdownMode}{Mode to choose from}

\item{includeDesc}{Include descriptions of data variables}
}
\value{
a tibble
}
\description{
Custom get function to query data from the UN Comtrade API.
}
\details{
further details on API features available at: `https://comtradedeveloper.un.org/api-details#api=comtrade-v1&operation=get-get`
}
\examples{
getUNComtrade(freqCode = "M",cmdCode = "010129" ,reporterCode = 36, startDate = c("2020-01-01"),endDate = "2020-02-01")
}
