% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscodes-data.R
\docType{data}
\name{hs_code}
\alias{hs_code}
\title{NZHSC codes and SOPI descriptions for comtrader shiny app}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3292 rows and 3 columns.
}
\usage{
hs_code
}
\description{
NZHSC codes and SOPI descriptions for comtrader shiny app
}
\keyword{datasets}
