% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hscodes}
\alias{hscodes}
\title{HS Codes}
\format{
A data frame with the following columns:
\describe{
  \item{sopiLevel}{SOPI level.}
  \item{sopiFilter}{SOPI level category.}
  \item{code}{Comma-separated string of NZHSC Level 6 codes by SOPI level and filter.}
}
}
\source{
V2022.5.0 \url{https://aria.stats.govt.nz/aria/#ClassificationView:uri=http://stats.govt.nz/cms/ClassificationVersion/GT9Qz2lTFnrbgpiI}
}
\usage{
hscodes
}
\description{
Dataset with HS and SOPI codes and descriptions for the ctdashboard shiny app.
}
\details{
hscodes

The `SOPI_group_HS6` category contains custom categories.
}
\keyword{datasets}
