% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscodeshiny.R
\docType{data}
\name{hscodeshiny}
\alias{hscodeshiny}
\title{hscodeshiny}
\format{
\describe{
  \item{sopiLevel}{SOPI level.}
  \item{sopiFilter}{SOPI level category.}
  \item{code}{comma separated string of NZHSC Level 6 codes by SOPI level and filter.}
}
}
\source{
NZHSC
}
\usage{
hscodeshiny
}
\description{
Dataset with NZHSC and SOPI codes and descriptions for comtrader shiny app.
}
\details{
The SOPI_group_HS6 category contains custom categories.
}
\keyword{datasets}
