% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSopiCheck.R
\name{hsSopiCheck}
\alias{hsSopiCheck}
\title{Identify conflicts between HS codes and SOPI categories}
\usage{
hsSopiCheck(ref1, ref2)
}
\arguments{
\item{ref1}{reference code 1 E.g.: NZHSC level code (either `2`, `4` or `6`)}

\item{ref2}{reference code 2 E.g.: SOPI category level (either `Primary Industry Sector`,`SOPI Forecast Group` or `Granular Group`)}
}
\value{
a tibble
}
\description{
Reads `omtcodes` and creates a wide table of HS codes and counts of matching SOPI Categories (or viceversa).
}
\examples{
hsSopiCheck(ref1 = `4`, ref2 = `SOPI Forecast Group`)
}
