% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs-code-data.R, R/omtcodes.R
\docType{data}
\name{omtcodes}
\alias{omtcodes}
\title{NZHSC codes and SOPI descriptions}
\format{
An object of class \code{data.frame} with 1646 rows and 6 columns.

\describe{
  \item{Primary Industry Sector}{SOPI Primary Industry Sector categories.}
  \item{NZHSCLevel4}{NZHSC Level 4 codes.}
  \item{NZHSCLevel4Desc}{NZHSC Level 4 descriptions.}
  \item{NZHSCLevel6}{NZHSC Level 6 codes.}
  \item{NZHSCLevel6Desc}{NZHSC Level 6 descriptions.}
  \item{SOPI_group_HS6}{SOPI group descriptions based on NZHSC Level 6 codes groups.}
}
}
\source{
NZHSC
}
\usage{
omtcodes

omtcodes
}
\description{
Dataset with NZHSC and SOPI codes and descriptions.
}
\details{
The SOPI_group_HS6 category contains custom categories.
}
\keyword{datasets}
