% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{reportercodes}
\alias{reportercodes}
\title{reporter codes}
\format{
a list
\describe{
  \item{List name}{country description}
  \item{List item}{country code}
}
}
\source{
https://comtradeapi.un.org/files/v1/app/reference/Reporters.json
}
\usage{
reportercodes
}
\description{
Named list of reporter country codes to use in comtrader shiny app
}
\details{
The list includes M49 codes for individual countries
}
\keyword{datasets}
