% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{set_uncomtrade_key}
\alias{set_uncomtrade_key}
\title{Support functions}
\usage{
set_uncomtrade_key(key)
}
\arguments{
\item{key}{The value to set as the key}
}
\description{
Custom support functions for `comtrader`
Set package key for UNComprade API
}
