% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSopiCheck.R
\name{cCheck}
\alias{cCheck}
\title{Identify conflicts between two categories in a table}
\usage{
cCheck(x = comtrader::omtcodes, ref1, ref2)
}
\arguments{
\item{x}{a data frame}

\item{ref1}{first reference column}

\item{ref2}{second reference column}
}
\value{
a tibble
}
\description{
Creates a table of counts of matching categories between two columns of a table
}
\details{
the table is built with reference to `ref1`
}
\examples{
cCheck(x = comtrader::omtcodes, ref1 = NZHSCLevel4, ref2 = `SOPI_group_HS6`)
}
