% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{commodityRecode}
\alias{commodityRecode}
\title{Create named vector for HS codes}
\usage{
commodityRecode(
  hsLevel = NZHSCLevel6,
  sopiLevel = SOPI_group_HS6,
  sopiFilter = NULL,
  query = TRUE
)
}
\arguments{
\item{hsLevel}{NZHSC level code (either NZHSCLevel2, NZHSCLevel4 or NZHSCLevel6)}

\item{sopiLevel}{SOPI category level (either `Primary_Industry_Sector` or `SOPI_group_HS6`)}

\item{sopiFilter}{a character vector indicating SOPI category level to filter}

\item{query}{If `TRUE` returns a vector of commodity codes (`cmdCode`) to use in a query. If `FALSE` returns a vector of `sopiLevel` names associated with the commodity code.}
}
\value{
a named character vector
}
\description{
Creates a named character vector of matching HS codes and SOPI Categories
}
\details{
Data was extracted from the Power BI OMT dataset. This dataset needs to be updated if there are changes in the OMT classification
}
\examples{
commodityRecode(hsLevel = NZHSCLevel6, sopiLevel = Primary_Industry_Sector, sopiFilter = "Dairy")
}
