% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCTSopi.R
\name{getCTSopi}
\alias{getCTSopi}
\title{Custom get data from UN Comtrade Database API}
\usage{
getCTSopi(
  typeCode = "C",
  freqCode = "M",
  clCode = "HS",
  reporterCode = NULL,
  startDate = NULL,
  endDate = NULL,
  partnerCode = NULL,
  partner2Code = NULL,
  flowCode = NULL,
  customsCode = NULL,
  motCode = NULL,
  aggregateBy = NULL,
  breakdownMode = "classic",
  includeDesc = TRUE,
  sopiLevel = SOPI_group_HS6,
  sopiFilter = NULL,
  hs = NZHSCLevel6
)
}
\arguments{
\item{typeCode}{Type of trade: C for commodities and S for service.}

\item{freqCode}{Trade frequency: A for annual and M for monthly.}

\item{clCode}{Trade (IMTS) classifications: HS, SITC, BEC or EBOPS.}

\item{reporterCode}{Reporter code (Possible values are M49 code of the countries).}

\item{startDate}{Optional start date for query (e.g. "2020-01-01"). Multi value input should be in the form of a character vector.}

\item{endDate}{Optional end date for query (e.g. "2022-01-01").}

\item{partnerCode}{Partner code (Possible values are M49 code of the countries).}

\item{partner2Code}{Second partner/consignment code (Possible values are M49 code of the countries).}

\item{flowCode}{Trade flow code. Multi value input should be in the form of a character vector).}

\item{customsCode}{Customs code. Multi value input should be in the form of a character vector).}

\item{motCode}{Mode of transport code. Multi value input should be in the form of a character vector).}

\item{aggregateBy}{Add parameters in the form of a character vector on which you want the results to be aggregated.}

\item{breakdownMode}{Mode to choose from.}

\item{includeDesc}{Include descriptions of data variables.}

\item{sopiLevel}{SOPI level column in `omtcodes` (either `Primary Industry Sector` or `SOPI_group_HS6`).}

\item{sopiFilter}{character string to filter SOPI level (e.g.: "Dairy").}

\item{hs}{HS code level column in `omtcodes`. E.g.: `Level 6`.}
}
\value{
a tibble
}
\description{
Custom get function to query data from the UN Comtrade API.
}
\details{
further details on API features available at: `https://comtradedeveloper.un.org/api-details#api=comtrade-v1&operation=get-get`
}
\examples{
getCTSopi(freqCode = "M", reporterCode = 36, startDate = "2020-01-01", endDate = "2020-02-01", sopiLevel = `Primary Industry Sector`, sopiFilter = "Dairy")
}
