% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{getDa}
\alias{getDa}
\title{Get dataset availability from UN Comtrade Database API}
\usage{
getDa(
  typeCode = "C",
  freqCode = "M",
  clCode = "HS",
  reporterCode = NULL,
  period = NULL,
  publishedDateFrom,
  publishedDateTo
)
}
\arguments{
\item{typeCode}{a character string indicating type of trade: "C" for commodities and "S" for service}

\item{freqCode}{a character string indicating trade frequency: "A" for annual and "M" for monthly}

\item{clCode}{a character string indicating trade classification (IMTS): "HS", "SITC", "BEC" or "EBOPS"}

\item{reporterCode}{a character string indicating reporter code (Possible values are M49 code of the countries)}

\item{period}{a character string indicating period. Year ("YYYY") or yearmonth ("YYYYMM"). Use character vector for multiple periods (defaults to all)}

\item{publishedDateFrom}{a character string indicating the start publication date ("YYYY-MM-DD")}

\item{publishedDateTo}{a character string indicating the end publication date ("YYYY-MM-DD")}
}
\value{
a tibble
}
\description{
Query datasets available in the UN Comtrade API
}
\details{
for further details visit the [UN Comtrade API developer site](`https://comtradedeveloper.un.org`)
}
\examples{
getDa(clCode = "H6", publishedDateFrom = "2020-01-01")
}
