% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{getLiveUpdate}
\alias{getLiveUpdate}
\title{Get live update data from UN Comtrade Database API}
\usage{
getLiveUpdate()
}
\value{
a tibble
}
\description{
Query progress on data release in the UN Comtrade API
}
\details{
for further details visit the [UN Comtrade API developer site](`https://comtradedeveloper.un.org`)
}
\examples{
getLiveUpdate()
}
