% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Get metadata from UN Comtrade Database API}
\usage{
getMetadata(typeCode = "C", freqCode = "M", clCode = "HS")
}
\arguments{
\item{typeCode}{a character string indicating type of trade: "C" for commodities and "S" for service}

\item{freqCode}{a character string indicating trade frequency: "A" for annual and "M" for monthly}

\item{clCode}{a character string indicating trade classification (IMTS): "HS", "SITC", "BEC" or "EBOPS"}
}
\value{
a tibble.
}
\description{
Query meta data from the UN Comtrade API
}
\details{
further details on API features available at: `https://comtradedeveloper.un.org/api-details#api=comtrade-v1`.
}
\examples{
getMetadata(clCode = "H6")
}
