% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTariffline.R
\name{getTariffline}
\alias{getTariffline}
\title{Get tariff line data from UN Comtrade Database API}
\usage{
getTariffline(
  typeCode = "C",
  freqCode = "M",
  clCode = "HS",
  reporterCode = NULL,
  period = NULL,
  partnerCode = NULL,
  partner2Code = NULL,
  cmdCode = NULL,
  flowCode = NULL,
  customsCode = NULL,
  motCode = NULL,
  includeDesc = TRUE
)
}
\arguments{
\item{typeCode}{a character string indicating type of trade: "C" for commodities and "S" for service}

\item{freqCode}{a character string indicating trade frequency: "A" for annual and "M" for monthly}

\item{clCode}{a character string indicating trade classification (IMTS): "HS", "SITC", "BEC" or "EBOPS"}

\item{reporterCode}{a character string indicating reporter code (Possible values are M49 code of the countries)}

\item{period}{a character string indicating period. Year ("YYYY") or yearmonth ("YYYYMM"). Use character vector for multiple periods (defaults to all)}

\item{partnerCode}{Partner code (M49 International Country Classification)}

\item{partner2Code}{Second partner/consignment code (M49 International Country Classification)}

\item{cmdCode}{a character string indicating HS commodity code. Use character vector for multiple commodity code entries (defaults to all)}

\item{flowCode}{a character string indicating trade flow code:  "X" for exports, "RX" for re-exports, "M" for imports, "RM" for re-imports. Use character vector for multiple trade flow entries}

\item{customsCode}{a character string indicating customs code. Use character vector for multiple customs code entries (defaults to all)}

\item{motCode}{a character string indicating mode of transport code. Use character vector for multiple customs code entries transport code (defaults to all)}

\item{includeDesc}{boolean indicating if categories descriptions shoould be returned (defaults to `TRUE`)}
}
\value{
a tibble
}
\description{
Query tariff line data in the UN Comtrade API
}
\details{
further details on API features available at: `https://comtradedeveloper.un.org/api-details#api=comtrade-v1`
}
\examples{
getTariffline(period = "202201")
}
