% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscodes.R
\docType{data}
\name{hscodes}
\alias{hscodes}
\title{hscodes}
\format{
\describe{
  \item{sopiLevel}{SOPI level.}
  \item{sopiFilter}{SOPI level category.}
  \item{code}{comma separated string of NZHSC Level 6 codes by SOPI level and filter.}
}
}
\source{
NZHSC
}
\description{
Dataset with HS and SOPI codes and descriptions for comtrader shiny app.
}
\details{
The SOPI_group_HS6 category contains custom categories.
}
