% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hscodes-data.R, R/hscodes.R
\docType{data}
\name{hscodes}
\alias{hscodes}
\title{NZHSC codes and SOPI descriptions for comtrader shiny app}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1700 rows and 3 columns.

\describe{
  \item{sopiLevel}{SOPI level.}
  \item{sopiFilter}{SOPI level category.}
  \item{code}{comma separated string of NZHSC Level 6 codes by SOPI level and filter.}
}
}
\source{
NZHSC
}
\usage{
hscodes
}
\description{
Dataset with HS and SOPI codes and descriptions for comtrader shiny app.
}
\details{
The SOPI_group_HS6 category contains custom categories.
}
\keyword{datasets}
