% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{hscodes}
\alias{hscodes}
\title{hs codes}
\format{
\describe{
  \item{sopiLevel}{SOPI level}
  \item{sopiFilter}{SOPI level category}
  \item{code}{comma separated string of NZHSC Level 6 codes by SOPI level and filter}
}
}
\source{
V2022.5.0 https://aria.stats.govt.nz/aria/#ClassificationView:uri=http://stats.govt.nz/cms/ClassificationVersion/GT9Qz2lTFnrbgpiI
}
\usage{
hscodes
}
\description{
Dataset with HS and SOPI codes and descriptions for comtrader shiny app
}
\details{
The SOPI_group_HS6 category contains custom categories
}
\keyword{datasets}
