% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{omtcodes}
\alias{omtcodes}
\title{omt codes}
\format{
\describe{
  \item{Primary Industry Sector}{SOPI Primary Industry Sector categories}
  \item{NZHSCLevel4}{NZHSC Level 4 codes}
  \item{NZHSCLevel4Desc}{NZHSC Level 4 descriptions}
  \item{NZHSCLevel6}{NZHSC Level 6 codes}
  \item{NZHSCLevel6Desc}{NZHSC Level 6 descriptions}
  \item{SOPI_group_HS6}{SOPI group descriptions based on NZHSC Level 6 codes groups}
}
}
\source{
NZHSC V2022.5.0 https://aria.stats.govt.nz/aria/#ClassificationView:uri=http://stats.govt.nz/cms/ClassificationVersion/GT9Qz2lTFnrbgpiI
}
\usage{
omtcodes
}
\description{
Dataset with NZHSC and SOPI codes and descriptions
}
\details{
The SOPI_group_HS6 category contains custom categories
}
\keyword{datasets}
