% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs-code-data.R, R/omtcodes.R
\docType{data}
\name{omtcodes}
\alias{omtcodes}
\title{Prepare data to match Comtrade and SOPI HS codes}
\format{
An object of class \code{data.frame} with 1646 rows and 6 columns.

a tibble with HS code levels 4 and 6 and matching SOPI category levels. The SOPI_group_HS6 category contains custom categories.
}
\source{
NZHSC
}
\usage{
omtcodes

omtcodes
}
\description{
Dataset with Comtrade and SOPI HS codes
}
\keyword{data}
\keyword{datasets}
