% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omtcodes}
\alias{omtcodes}
\title{OMT Codes}
\format{
A data frame with the following columns:
\describe{
  \item{Primary_Industry_Sector}{SOPI Primary Industry Sector categories.}
  \item{NZHSCLevel4}{NZHSC Level 4 codes.}
  \item{NZHSCLevel4Desc}{Descriptions for NZHSC Level 4 codes.}
  \item{NZHSCLevel6}{NZHSC Level 6 codes.}
  \item{NZHSCLevel6Desc}{Descriptions for NZHSC Level 6 codes.}
  \item{SOPI_group_HS6}{SOPI group descriptions based on NZHSC Level 6 code groups.}
}
}
\source{
NZHSC V2022.5.0 \url{https://aria.stats.govt.nz/aria/#ClassificationView:uri=http://stats.govt.nz/cms/ClassificationVersion/GT9Qz2lTFnrbgpiI}
}
\usage{
omtcodes
}
\description{
Dataset with NZHSC and SOPI codes and descriptions.
}
\details{
omtcodes

The `SOPI_group_HS6` category contains custom categories.
}
\keyword{datasets}
