% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partnercodes}
\alias{partnercodes}
\title{Partner Codes}
\format{
A named list where:
\describe{
  \item{List name}{Country description.}
  \item{List item}{Country code.}
}
}
\source{
\url{https://comtradeapi.un.org/files/v1/app/reference/partnerAreas.json}
}
\usage{
partnercodes
}
\description{
Named list of partner country codes to use in the ctdashboard shiny app.
}
\details{
partnercodes

The list includes a "World" code in addition to M49 codes for individual countries.
}
\keyword{datasets}
