% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partnercodes-data.R, R/partnercodes.R
\docType{data}
\name{partnercodes}
\alias{partnercodes}
\title{Partner country codes and descriptions}
\format{
An object of class \code{list} of length 310.

A list.
\describe{
  \item{List name}{country description.}
  \item{List item}{country code}
}
}
\source{
https://comtradeapi.un.org/files/v1/app/reference/partnerAreas.json
}
\usage{
partnercodes
}
\description{
Named list of partner country codes to use in comtrader shiny app.
}
\details{
The list includes a "World" code in addition to M49 codes for individual countries.
}
\keyword{datasets}
