% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reportercodes}
\alias{reportercodes}
\title{Reporter Codes}
\format{
A named list where:
\describe{
  \item{List name}{Country description.}
  \item{List item}{Country code.}
}
}
\source{
\url{https://comtradeapi.un.org/files/v1/app/reference/Reporters.json}
}
\usage{
reportercodes
}
\description{
Named list of reporter country codes to use in the ctdashboard shiny app.
}
\details{
reportercodes

The list includes M49 codes for individual countries.
}
\keyword{datasets}
