% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportercodes-data.R, R/reportercodes.R
\docType{data}
\name{reportercodes}
\alias{reportercodes}
\title{Reporter country codes and descriptions}
\format{
An object of class \code{list} of length 255.

A list.
\describe{
  \item{List name}{country description.}
  \item{List item}{country code}
}
}
\source{
https://comtradeapi.un.org/files/v1/app/reference/Reporters.json
}
\usage{
reportercodes
}
\description{
Named list of reporter country codes to use in comtrader shiny app.
}
\details{
The list includes M49 codes for individual countries.
}
\keyword{datasets}
