% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_uncomtrade_key.R
\name{set_uncomtrade_key}
\alias{set_uncomtrade_key}
\title{Set package key for UNComprade API}
\usage{
set_uncomtrade_key(key)
}
\arguments{
\item{key}{The value to set as the key}
}
\description{
Set package key for UNComprade API
}
